+++ binstar_client/commands/login.py
@@ -1,45 +1,39 @@
-'''
+"""
 Authenticate a user
-'''
+"""
 from __future__ import unicode_literals
 
 import getpass
 import logging
+import platform
 import socket
 import sys
-try:
-    from urlparse import urlparse
-except ImportError:
-    from urllib.parse import urlparse
+
+from six.moves.urllib.parse import urlparse
+from six.moves import input
+
 from binstar_client import errors
-from binstar_client.utils import get_config, get_server_api, store_token, \
-    bool_input
-import platform
+from binstar_client.utils import get_config, get_server_api, store_token, bool_input
 
 
 logger = logging.getLogger('binstar.login')
 
-try:
-    input = raw_input
-except NameError:
-    input = input
-
 
 def try_replace_token(authenticate, **kwargs):
-    '''
+    """
     Authenticates using the given *authenticate*, retrying if the token needs
     to be replaced.
-    '''
+    """
 
     try:
         return authenticate(**kwargs)
     except errors.BinstarError as err:
         if kwargs.get('fail_if_already_exists') and len(err.args) > 1 and err.args[1] == 400:
             logger.warning('It appears you are already logged in from host %s' % socket.gethostname())
-            logger.warning('Logging in again will remove the previous token. '
-                     ' (This could cause troubles with virtual machines with the same hostname)')
-            logger.warning('Otherwise you can login again and specify a '
-                      'different hostname with "--hostname"')
+            logger.warning('Logging in again will remove the previous token. (This could cause troubles with virtual '
+                           'machines with the same hostname)')
+            logger.warning('Otherwise you can login again and specify a different hostname with "--hostname"')
+
             if bool_input("Would you like to continue"):
                 kwargs['fail_if_already_exists'] = False
                 return authenticate(**kwargs)
@@ -49,7 +43,7 @@ def try_replace_token(authenticate, **kwargs):
 
 def interactive_get_token(args, fail_if_already_exists=True):
     bs = get_server_api(args.token, args.site)
-    config = get_config(remote_site=args.site)
+    config = get_config(site=args.site)
 
     token = None
     # This function could be called from a totally different CLI, so we don't
@@ -59,7 +53,7 @@ def interactive_get_token(args, fail_if_already_exists=True):
     url = config.get('url', 'https://api.anaconda.org')
 
     auth_name = 'binstar_client:'
-    if site and site != 'binstar':
+    if site and site not in ('binstar', 'anaconda'):
         # For testing with binstar alpha site
         auth_name += '%s:' % site
 
@@ -82,9 +76,7 @@ def interactive_get_token(args, fail_if_already_exists=True):
             raise errors.BinstarError(
                 'Unable to authenticate via Kerberos. Try refreshing your '
                 'authentication using `kinit`')
-
     else:
-
         if getattr(args, 'login_username', None):
             username = args.login_username
         else:
@@ -116,7 +108,6 @@ def interactive_get_token(args, fail_if_already_exists=True):
                 password = None
                 continue
 
-
         if token is None:
             parsed_url = urlparse(url)
             if parsed_url.netloc.startswith('api.anaconda.org'):
@@ -131,30 +122,23 @@ def interactive_get_token(args, fail_if_already_exists=True):
 
     return token
 
+
 def interactive_login(args):
     token = interactive_get_token(args)
     store_token(token, args)
     logger.info('login successful')
 
+
 def main(args):
     interactive_login(args)
 
+
 def add_parser(subparsers):
-    subparser = subparsers.add_parser('login',
-                                      help='Authenticate a user',
-                                      description=__doc__)
+    subparser = subparsers.add_parser('login', help='Authenticate a user', description=__doc__)
     subparser.add_argument('--hostname', default=platform.node(),
-                           help="Specify the host name of this login, "
-                                "this should be unique (default: %(default)s)"
-                           )
-    subparser.add_argument('--username',
-                           dest='login_username',
-                           help="Specify your username. "
-                                "If this is not given, you will be prompted"
-                           )
-    subparser.add_argument('--password',
-                           dest='login_password',
-                           help="Specify your password. "
-                                "If this is not given, you will be prompted"
-                           )
+                           help="Specify the host name of this login, this should be unique (default: %(default)s)")
+    subparser.add_argument('--username', dest='login_username',
+                           help="Specify your username. If this is not given, you will be prompted")
+    subparser.add_argument('--password', dest='login_password',
+                           help="Specify your password. If this is not given, you will be prompted")
     subparser.set_defaults(main=main)