+++ binstar_client/scripts/cli.py
@@ -45,7 +45,7 @@ def file_or_token(value):
 
 def _custom_excepthook(logger, show_traceback=False):
     def excepthook(exc_type, exc_value, exc_traceback):
-        if issubclass(exc_type, KeyboardInterrupt) or not issubclass(exc_type, errors.NotFound):
+        if issubclass(exc_type, KeyboardInterrupt) or not issubclass(exc_type, errors.ServerError):
             return
 
         if show_traceback: