+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Added
 
+* Added support for setting a default upload user
 * Show warnings when lock down and read-only headers are present
 
 ## Version 1.6.9 (2018/01/31)